Consolidated image-loading code, made it async, misc improvements.
The slideshow view is now stored in  the renamed SingleImageViewModel. Displaying any image should go through a new VM, ImageViewModel, which handles loading image data of one image. It supports asynchronicity and initial loading of the image at a preview resolution. With this approach, XAML image tags should directly bind to the ImageSource stored in ImageViewModel instead of going through a converter. The converters have been removed.

NavigatorViewModel now calls a new ViewModelBase virtual function when a VM page is removed. This function is so the VM page can cleanup after itself (in the case of ImageViewModel, cancel any async image loading tasks).