Added caching of nearby images in slideshow mode.
Now, when viewing an image full-size in slideshow mode, the program will load the nearby images (right now, one before and two after) in the background, so that hopefully the user will have to wait less for images to load. This particularly works well if the user pauses to look at each image.