In slideshow view, fixed incomplete list of photos and improved loading.
Now the entire gallery of ImageViewModels is initialized first and the images themselves loaded after. This way, the gallery doesn't extend in length over time: it's fixed length, and the images just load in & render over time. This also means that the slideshow view will span over the whole gallery instead of just the already-loaded images.

In slideshow view, when the view is moved to the right or left, changed where the display update is called so that if the current image hasn't finished loading, the outdated image in that cache position doesn't display.