Now loads image lists before loading images.
Before, the gallery's ImageViewModel objects would load their images directly after they were added to the gallery, which meant that the gallery's length was increasing as each image loaded. Now, it intializes all the ImageViewModel objects and adds them to the gallery first, and then loads each image. This way, the gallery's length is fixed from the beginning, and the images load into those spaces over time.

This is helpful because the slideshow view takes a list of ImageViewModels from the gallery as its source, so if you were to click on an image in the gallery before all the images were loaded, the slideshow view would be limited to viewing only the images that were loaded at the time of its creation. With this change, you can view all of the images in the gallery, whether or not their data was loaded when you went into slideshow mode.