Now the gallery view loads images in the user's view first.
Previously, the gallery would load the image thumbnails one at a time, from start to finish. Now, every time the user scrolls to an area of the gallery, the program will first load the images that are in view, and then go back to loading the rest of the gallery sequentially.

A new Utils folder was added with a class containing utility functions related to displaying content to the user.

There is a new dependency (System.Windows.Interactivity) that allows commands to be called when an event is triggered. This let's me follow the MVVM pattern by avoiding event-handling code in the code behind.