Drop-down chooser control, & defocus on off click.
Created a separate UserControl for a filterable drop-down list, in ChooserDropDownViewModel & ChooserDropDownView. Typing in a textbox pulls up a drop-down list of items which are filtered by what is in the textbox. Choosing an item in that drop-down (or pressing the separate button "Create new") will call a callback function provided by the user of the Control. Adding tags in the ImageInfo panel and choosing filtering tags in the Gallery view have both been switched to using this Control.

A click event handler has been added to the Window, which will be triggered whenever the user clicks on something non-focusable like a panel, grid, or label. This handler will set the focus & keyboard to the window's main grid, taking focus from whatever had it before. Basically, the user can click off of a textbox, for example, and have it lose focus.

Added an extended ObservableCollection class, RangeObservableCollection, that provides batch operations on the collection with only one notification send to observers.