Added support for video files along with images.
Photo and PhotoGallery have been renamed to Media and MediaGallery and support both image and video. The ViewModels have been given a base class MediaViewModel and the subclasses are ImageViewModel and VideoViewModel. ImageSlideshowViewModel was renamed to SlideshowViewModel, and ImageInfoViewModel was renamed to MediaInfoViewModel.

Right now, on the gallery view, the videos are loaded but paused for each thumbnail, which is not efficient at all. In the future, it will need to load thumbnails for each video.

Right now, in slideshow view, videos will play automatically. In the future, controls will be added to pause/play the videos.

Fixed an issue in RangeObservableCollection