Fixed image cache issues with small gallery sizes.
The caching system crashed when given fewer items than cache positions. The new system will shrink the cache if needed to fit the number of items. In this case, it will not trigger any image reloads when the user moves between images because none of the images will ever be evicted from the cache.