Videos are now displayed in the gallery as thumbnails.
The Video model class now stores an Image model instance for it's thumbnail. Now, when creating an VideoViewModel, the user can choose to put it in Thumbnail Mode. In Thumbnail Mode, the VideoViewModel will create & maintain an ImageViewModel for the video's thumbnail.

The gallery view has its VideoViewModels in Thumbnail Mode and when rendering them, renders the thumbnails like it does any other image, instead of loading the whole video into memory and pausing on the first frame. This is a huge memory saver.