Videos now use existing thumbs instead of making new ones
Previously, videos would generate a new thumbnail every run. Now, saving the gallery to xml will save the thumbnail path, and the video will not regenerate a thumbnail if it has a path saved in the xml file. To do this, the thumbnail loading code was put in a separate function that is only called by those who need it. In VideoViewModel, if Thumbnail Mode is enabled in the constructor, it will call that function. That function will only generate a thumbnail if one does not already exist for that video.