Now checks image existance on load.
After loading in the MediaGallery instance from the xml file, the program checks that all the media in the gallery still exists in the filesystem. Currently, if a media item does not exist, it will be removed from the gallery.

Does not explicitly check that video thumbnails exist, because that is done when creating the VideoViewModel (by calling the Video object's LoadThumbnail() method).