Made ChooserDropDown much more functional.
A big overhaul to the ChooserDropDown control.

The "Create New" button is no longer a button that is
styled separately from the rest of the item list. It's now included by the XAML as the last item in the list. This means that it's processed as just another tag, so that when it's processing a tag selection, the view model must add a check for the text "Create New".

The up and down arrow keys scroll through the list. There will always be a selected item in the list (if the user doesn't change it with the arrow keys, it will be the first in the list). Pressing tab will autofill the textbox to the text of the selected item. Pressing enter will "choose" the selected item, whether that means choosing an actual item or creating a new one (when the selected item is the Create New button).

Added utility HandlingEventTrigger. It's a subclass of EventTrigger, which allows routing XAML events to ViewModel commands. Unlike the standard EventTrigger, which allows events to keep propogating, this one marks
the event as handled so that it stops propogating. It's not currently used, but I keep thinking I'm going to need it, so I'll leave it for now.