Changed ChooserDropDown "tab" autofill behavior.
Before, pressing Tab would autofill the ChooserDropDown textbox to the text of the currently selected item. Now, it behaves more like autofill on a command line would: it autofills the text up to the point where the items have the
same beginning substring. So if your strings were "abcd", "abcz", and "abcdef", and your input was "a", hitting Tab would autofill to "abc".