Made ChooserDropDown an independent UserControl.
Rewrote ChooserDropDown so that most of it's functionality was inside it as a UserControl, and removed it's attached ViewModel. Most of the logic didn't need to be in a ViewModel, and having it there made it slightly complicated to hook up. So now it's much nicer to plop a ChooserDropDown in your XAML, without having to keep a VM reference somewhere.