Dynamically adjusts image height for smooth rows
Similar to how Google Photos works, the images in a row are made to be the same height, and to adjust their height so that each row is fully filled. This works decently, but the right sides of the last images can be a few pixels off. I believe this is due to padding between the images, but I don't have a great way to determine the padding value right now.