Added group tag button, fixed tag update functions
Added "Add Tag To Selected" button. This is a ChooserDropDown that lets the user add a tag to all of the currently selected images in the gallery.

MediaViewModel has a property "IsSelected", whether it's selected by its parent collection. Each collection should have its own MediaViewModel instance, so this is fine.

UpdateTags() will now only add tags (send add notifications) for tags that weren't previously in the collection, and it will remove any obsolete tags.

Removed RangeObservableCollection because it wasn't working as expected. It turns out the notification doesn't just say "check the list", it actually passes the values that have changed. So I was triggering the "add" notification with only one item, when really it should be "add" with a range of items. Added RemoveRange and removed the other function.