Fixed binding error bug adding a tag to multiple media.
When adding a tag to multiple media with the AddTagToSelected button, it would call Refresh() on the ICollectionView ImageView multiple times. This was causing binding errors on MinHeight and ContentAlignment properties. These are apparently harmless errors that are handled internally, but it's annoying that they're there, and they could mask real issues.

So to fix this, I added a "DisableTagUpdate" property in MediaGallery. This lets me do batch operations with tags, without updating the list in their associated MediaGallery. I manually call UpdateTags in MediaGallery after the operation is done, and that works fine.