Added shift-click media selection and reworked focus.
Added a shift-click feature on the gallery media select buttons. The previously selected item is stored, and shift-click select will select or deselect all the items inbetween the current and previous selection.

Pressing escape while in the gallery will deselect any media that are selected.

ChooserDropDown plugins now close their popups when escape is pressed. Previously, closing the popup passed focus back to the main window, but now it passes it to the nearest focusable parent.

Reworked the focusing system. Previously, the main window always held focus and it passed its key events down to the navigator, which passed them to the current page. It's a lot simpler to just have the focus be on the current page directly. For this to work when switching pages, each page must be Focusable and claim focus on its Loaded event.

For clicking on the background of a page to claim focus, mouse clicks were previously captured in the MainWindow. They're now captured on a page-by-page basis. Note that the background must not be None for mouse events to be captured. I've set the page backgrounds to Transparent and this works well.

If ChooserDropDowns are open when the user switches to another application, their state will be restored when the user switches back, and focus will be on the ChooserDropDown again. I struggled with this for a long time. I could get the popup to display, but on the window Activated event handler, focusing on anything in the popup would simply fail. I eventually discovered that first focusing on something outside of the popup (like the MainWindow) and then focusing on the popup. I'm somewhat clueless as to why this is, but it works, so I'm going to roll with it.