Separated out the media display from the gallery page. Now, a set of Media items is called a Collection, and the Gallery page is a specific way of viewing and interacting with that collection.
The model MediaGallery is now called MediaCollection (and it has its own MediaCollectionViewModel), and the view for displaying the list of Media is now separated into the MediaCollection view. The Gallery view and ViewModel now contain the MediaCollection and view model. The logic for displaying the images, etc. is contained in MediaCollection and its ViewModel, and Gallery and its view model contain only the filtering and controls at the top.

Moved MediaViewModel generator to a static generator in MediaViewModel. It's used by MediaCollectionViewModel and GalleryViewModel, but there's a version in SlideshowViewModel that doesn't fit that layout (because it needs to insert an event handler on the image only). So I'll keep it for now, but I'm not fully decided.

Attempted to organize ChooserDropDown by adding macro regions.