Updated RangeObservableCollection event, documented
RangeObservableCollection now sends a Reset event instead of an Add/Remove event. Before, it was using an Add or Remove event (which handles the adding/removing of one item) after changing a whole range of things. The Reset event is more appropriate, because it means the collection has "changed dramatically".