Moved some event handler, collection and tag change events, to MediaCollectionViewModel from GalleryViewModel, because they should happen regardless of whether a gallery is attached to the MediaCollectionVM.
For whatever reason, removing Media from the MediaCollection and updating the ICollectionView MediaView implicitly calls the Refresh() function. Thus, if Refresh() is called in MediaTagsChanged, binding errors occur. So we need to disable Refresh() when removing media. Added a bool in MediaCollectionViewModel to do that.

Thought about getting rid of the MediaCollection DisableTagUpdate functionality, because it could be done by disabling Refresh() with the MediaCollectionVM bool instead. But the issue with that is that only the MediaCollectionVM can disable those calls. MediaCollection shouldn't have any knowledge of MediaCollectionVM. If another VM were to update the tags, MediaCollectionVM would stlil have multiple updates. So it should stay in MediaCollection.

Renamed the RemoveTag command and associated variables to RemoveTagFromFilter to be more consistent.