Created basic sidebar navigation, publicized classes.
Created a prototype sidebar for switching between pages. The details of how pages are related is still to be worked out. For example, do the maps and gallery pages function as separate in the view of the navigator, or can you go back from one to the other.

Also had to make a couple viewmodel classes public.