Added initial map prototyping
Using the Bing Map control, the map is displayed in the Map view and associated with the MapViewModel. The model MapLocation has a MapLocationViewModel. These locations are displayed on the map using the Bing Pushpin control. If you click on the Pushpin, it will open a popup on the map called the MapLocationPreview view.