Added Timestamp to all Media, sorted gallery
Added a Timestamp property to all Media. The Gallery view now sorts by timestamp. The MediaInfo tab now shows the timestamp.

On load, the timestamp of the media is determined. For all kinds of media, this means looking at the FileCreated and FileModified metadata, as well as seeing if the filename matches the date format I use (yyyymmdd__hhmmss.ext). For images, it also considers the DateTaken field in the exif metadata if it exists. If the filename format provides a timestamp, it will pick that one. Otherwise, it will pick the earliest of all the collected dates. The idea is that if someone has named the filename something, that is an intentional choice and should override any other metadata options.

Fixed a bug where removing images from a MediaCollection would stop the filter from updating when media tags changed.