Added Events and an event page, added Gallery model.
Events are occurances over some time range that have media associated with them. This could be a week-long vacation, or a dinner party. The time-range part is not implemented yet.

Revised how Gallery and MediaCollection interact. The Gallery model object is now the top-level element - it contains all of a user's media, events, etc. It is what's saved in the xml database.

Created an EventsView and associated EventsViewModel to display a list of all the events. Not sure if this will stay in the future. EventViewModel objects are associated with specific Event objects. From the EventsView, you can click on a specific event and view that EventPage, which contains the name and a display of that event's MediaCollection.

Added a ChooserDropDown in the gallery view to add the selected items to an event.

Removed the Name field from MediaCollection.