Allowed events nested in MediaCollections.
Now, MediaCollections are a list of ICollectable objects, which is an abstract class that has both Media and Event as subclasses. The result is, a MediaCollection can contain an event, and an event has its own MediaCollection, so you can nest MediaCollections. ICollectable has its own VM too.

Events show up in MediaCollections, but they are not selectable. If you click on one, it will open the event's page.

Had to do a little bit of reworking to handle this, such as having to pass Tag list change listeners all the way down to the nested media.

Moved the MediaCollection Select button to individual data templates, rather than the global ItemTemplate, so that some items (Events) can not have select buttons.
Moved select button to datatemplate