Fixed bug where adding files caused binding errors
Was having an issue where adding items to a MediaCollection was calling Refresh(), when Refresh() was already getting called by adding the Media to the ObservableCollection the view is attached to.

So edited AddMediaItem() and added a method AddMediaItems() to MediaCollectionVM which adds multiple Media items at once. Both disable the call to Refresh() when the collection's tags update, because then adding a new VM to the MediaView will trigger a Refresh().