Added timestamp to sort events, AddToEvent button
Added a StartTimestamp property to Event which is the earliest timestamp of the media in the Event. The gallery sorts events using this.

Implemented the AddSelectedToEvent button. To do this, gallery contains a list of EventViewModels for all events nested in the gallery. When media is added to an event, it is removed from the gallery.

Added RemoveMediaItem(s) toMediaCollectionViewModel. Once at the end of these, and AddMediaItem(s), Refresh is called. This is not necessary to show the added items, but is necessary to resort the items.

Removed event list from the gallery model, because they're now stored in MediaCollections.

For some reason, pushing a new page with the existing EventViewModel will not display any media recently added to the event. Creating a new vm does work. Which is okay, because soon I'm going to separate the event page vm and the event tile with the MediaCollection view.

Along these lines, each EventViewModel stored in the _mediaList in MediaCollectionVM had its own MediaCollectionVM and was loading that media too. I disabled this in a hacky way, and this will be fixed in the VM reorganization.