Seperate VM for event page and event in collection.
Created an EventTileViewModel, which is used for displaying an event in the MediaCollection view. The main difference is that the EventTileViewModel does not contain a MediaCollectionViewModel and does not load any of the items in the MediaCollection. When an Event page is created, a new EventViewModel is created and that does load the images in the event.