Changeable thumbnail height & collection preview
There's now a button in the gallery which switches the thumbnail height between two values. Later this can have better customizability, but right now, the gallery will respond.

Also added a preview mode and visible row number to the MediaCollection display. There is a dependency property RowNum which determines how many rows of media are shown. If this is set to 0, functionality is normal. If it's set to a positive integer, then the height of the component is set to only show that number of rows. The scroll bar is also disabled.

There is a PreviewMode dependency property as well, but this doesn't do anything as of yet.