Consolidated MediaColl, added See All in event page.
Moved all the options buttons back into the MediaCollection control (except for Add Files and Save Gallery). Yes, the vm is kinda large, but it makes sense for it to all be together. Especially since I'm considering having buttons disappear when they're not enabled, it makes sense UI-wise.

Changed GalleryVM to take a MediaCollection model instead of a Gallery model. This is because I've added a See All button on event pages that opens a Gallery display of only the event's media.

In MediaCollectionVM, there was an extra refresh going on when adding items was causing binding errors. Turns out manually calilng refresh to resort the items only was needed when removing items.

Moved Save Gallery functionality to MainWindow.xaml.cs. Eventually the XML stuff will be in its own class.