Editable event name, allow creating new events
Created a control EditableLabel, which is a Label which can be edited by double-clicking. The label is replaced with a TextBox.

Made the event page seize focus on click, and when escape is pressed (which leaves edit mode for an open EditableLabel).

AddSelectedToEvent button now allows creating a new Event.

Previously, MediaCollectionVm had a bool to disable refresh of the MediaView. The issue with this was that nested functions would be disabling and enabling this and it would get kind of messy. So now, the variable is an int which keeps track of how many callers have disabled refreshing, and only allows refresh once all of them have reenabled it.