CreateEvent button, default thumbnails
Added Create New Event button which creates a blank event.

When an event is created, or an empty event is given media, chooses the first available Image model as the thumbnail.

The EventTileViewModel constructor will now create a VM for the thumbnail if the thumbnail exists. Failing to do so was causing aspect ratio issues. If some action later changes the thumbnail, it will need to call CreateThumbnailVM() to refresh the vm. It would be done by LoadThumbnail() when the MediaCollection loads its items, but the aspect ratio needs to be determined before that.