Added labels to separate media display by year & month, refactored, added custom MediaView.
Made a custom class acting like ICollectionView to manage the list of ICollectableViewModels in MediaCollectionViewModel. It does all the filtering, sorting, and creation of the VMs from the ICollectable models.

Part of the benefit of this is the ability to add labels to the MediaCollection display to separate the items by time range (year, month). MediaView will add TimeLabelViewModel instances into the list of items. In the MediaCollection control, it will allocate the width of the container to each of these, so they have their own row.

In general, followed a much better EventHandling flow. Both in terms of where notifications are generated and how they're processed.

For many event handlers, they process by action, so that if an item was added to the MediaCollection, for example, you can add the item into the MediaView list instead of rebuilding the whole list. Examples of use: Updating event time range when its collection items change; Updating the list of tags in a MediaCollection when an item is added, removed, or tags changed.

Moved the INotifyPropertyChanged implemention to a util class called NotifyPropertyChanged, so it could be used by models as well. Specifically, ICollectable subclasses it.

As more ViewModels access the same set of models, more models are going to need to notify ViewModels of their changes. ObservableCollections are nice because they already implement this.

When Events' StartTimestamp or Thumbnail change, they will raise PropertyChanged events. These are picked up by the EventTileViewModel. This way, as soon as a Thumbnail is chosen for an event, the VM can create an ImageVM for it. This also allows the MediaView to reorder its contents when events' time ranges change.

Media has a TagsChanged event for when the Tags collection changes. This is so that it can put itself as the sender and the event handlers can know which Media object's tags changed.

Similarly, MediaCollection has a MediaTagsChanged event which passes on any children CollectionChanged events.

Made it so tags within an event are not known by the MediaCollection containing the event. Is this what we want?

Added a Timestamp property to ICollectableViewModel to consolidate its different sources in the subclasses. To set this, subclasses override a protected getter method which the property uses.

Added method in RangeObservableCollection to clear the list and add several items, all with one reset notification.

Made events' time range properties null by default if the event has no content. The view models then detect this and return a default timestamp.

Moved the TimeRange enum to its own file.