Created Events page folder system
The Events page now displays a folder system with expandable folders. The top level is each year, the second level is each month, and subsequent levels are events within each month. Individual media are not shown.

If you click on an event folder, that event's collection will open up. If you click on a year or month folder, it will open up a view of all the media within that time range.

To do this, created the FolderLabelViewModel. It has no model behind it, because it is only used for display. It contains a label, a timestamp, and a collection of sub-folders.

Added PrecisionDateTime class to mark up to what precision a DateTime is defined. Switched every DateTime over to PrecisionDateTime, but kept the Media timestamps at Second precision always. Otherwise, there's going to have to be more checking in methods that use and compare those timestamps.