Selective display of TimeLabels in MediaView
MediaView now has a maxLabel argument in the constructor that determines the highest TimeRange label to display. If this is null, then it will display only TimeRanges for which there are multiple labels. This is used for events, so if an event takes place over multiple months, it will show the month labels. Otherwise not.