Misc updates, bug fix
Added View All Media button in the events page, made it the default page, and removed the Gallery option from the sidebar.

Fixed bug in RecalcImageSizes() where if the first image in the row was wider than the row, it would infinitely loop.

Made MediaView always show day labels, even when dynamic label showing is on.