Fixed bugs, added popup to pick event
Fixed bugs involving opening galleries for years or months. Previously, it was generating a new MediaCollection with the items, but this allowed the new MediaCollection and the original one to get out of sync. So now, it uses the original MediaCollection, and simply adds a filter to the gallery's MediaView that only shows media of the proper year or month. This feels like exactly what MediaView is meant to do.

Gave PopupWindow a dependency property bool called Open. It's bound to the ViewModel, and when its set to false, the popup will close. Thus, the ViewModel can close the popup, but it's still MVVM because it doesn't know anything about the DependencyProperty.

Added cleanup functionality to PopupViewModel so that the popup can do anything it needs to when it's closed, such as removing event handlers, etc. Note that it uses a function called Cleanup() instead of the deconstructor because there is no guarantee when garbage collection will happen. Added a cleanup function to the FolderView as well b/c that's used in the popup.

Created the EventSelectionPopup, which displays a folder view of the events in the collection and lets the user choose one. This is used to add items to an event. Along with this, created an EventSelectionPopupReturnArgs to pass data from the popup back to the main program.

Added option to FolderView that determines whether folders with no events are shown or not.

Also added an event handler to FolderView that passes along FolderOpened events to outside classes.

Made MediaView properly handle multiple Filter functions.