CreateNewEvent button, page cleanups when closed
Added Create New Event button to the Event Selection popup. This will open a TextEntryPopup, which just asks the user for a string, and use that name to create a new event and add the selected media to it.

Made MediaCollection only load items that are currently in the MediaView. Before it was loading the entire list of items, which is particularly bad now that there are some filters that only show one month's worth of media. No need to load all the other media then.

Added cleanup functions to every ViewModel and View that needed it, and made navigator call Cleanup whenever a page was removed.