Improved MediaView's efficiency & features
Previously, MediaView maintained a list of ICollectableViewModels for every single item in its collection, whether or not that item was visible. Now, it only creates ICollectableViewModels for visible items, which makes things a lot more simple and memory efficient. This means the Filtering functions now work on ICollectables instead of ICollectableVMs

MediaView now listens for changes in items' Timestamps and resorts when that happens.

Added an ExpandEvents mode to MediaView where events will be replaced with all of their child Media items.

Tried to make RangeObservableCollection work.  Turns out it works, except when it's bound to a WPF UI item, so it's simply not going to work. Back to a whole load of notifications..

Seperated out MediaView's item adding into an AddItem() method.