Revised filtering/search system, small class upgrades
Replaced tag filtering with a search page. This page will have options for filtering with all sorts of parameters (tag, time, location, etc.). You click "search", and then go to a results page where the relevant Media is displayed.

As part of this, overhauled the filtering system. Now, the MediaView is given a FilterSet instance, which is where the Filter() function and all the associated storage is done. The FilterSet contains several FilterCriteria instances, each of which contains filtering info for a different property. These are stored in different subclasses of FilterCriteria. Currently, TagFilter and TimeRangeFilter exist.

This is a nice division of labor. Now to view a certain month, you just pass along a FilterSet with the TimeRangeFilter set to that month when creating the GalleryViewModel.

Made it so that tags within an Event will show up in a parent MediaCollection's list of tags.

Made PrecisionDateTime <= and >= use Matches() instead of ==(). This way, users can use <= and >= to see if a Media item is within a time range of a certain precision. Perhaps ==() should just do what Matches() does.

Made ViewModelBase abstract in the hopes that I won't forget to implement Cleanup() as often.