Separated out UI elements & ViewModels of Search Page.
Cleaned up the Search Page a bit. Separated the seperate types of FilterCriteria out into seperate UI elements with separate ViewModels as their DataContexts. SearchPageViewModel now just has to expose properties for each FilterCriteriaViewModel. This is done in a hacky way because of how the FilterCriteria lazy loading is done in FilterSet, but, oh well, it's nice to have iteration through all the criteria.

Moved Search files to seperate folders.