Added a time range search criteria.
Added a search criteria for time range in the search page. Currently, the user can select the start and end time down to Day precision, but in the future, I would like more detailed selection.

When only a time range is selected in the search criteria, events will show up as condensed. If a tag is selected in the search criteria, then the events will be expanded and only the items with the tag in the event will be displayed.

The DatePicker control uses a DateTime object, so there is now a PrecisionDateTimeToDateTimeConverter that increases the precision in the proper way depending of if it's the start or end timestamp.

When increasing the precision of a PrecisionDateTime, the constructor now has the option to generate the missing data from the beginning of the precision range or the end. So if converting precision from year to month, the new timestamp's month will either be January or December.