Map paths & locations have previews, are editable
Maps now have two types of items that are displayed on them: Paths and Locations. Locations are single points, Paths are ordered collections of points.

The Map model contains a list MapItems, which are either MapPaths or MapLocations. The ViewModel and View classes follow the same structure.

Each type of map item has a Preview box, which will appear as a popup on the map when the item is clicked on. This is meant to be a quick view at the item, and should have a link to go to the full details of the item.

Each type of map item can be edited by clicking the Edit button in its preview. When the Map control enters Edit Mode, no previews can be shown and no items can be interacted with except for the currently edited item. When a location is edited, the user can drag its pushpin around. When a path is edited, pushpins appear for each point, which the user can drag around, click on to remove. The user can also click on the map to append a new point to the path, or click on the path itself to insert a point. When an item is being edited, a panel appears to the right side which has the details of the item (its name, all its points, etc.). The Map control toolbar has a button to take the item out of edit mode. Only one item can be in edit mode at a time.

Added utility class to find UIElement ancestors of a certain type from the code behind.

MapItemView and MapItemPreviewView are both abstract UserControl base classes. When making these, you cannot have an XAML part of the class, you must only create a C# file.