Overhauled the map path edit mode, drag behavior
The previous edit mode for Map Paths could not handle paths with thousands of points, as it was creating a Pushpin for each point. So instead, now there's a system where if the user's mouse is within a certain distance of a point, the closest point will have a Pushpin appear for it. If the user clicks on the Polyline, that Nearby Pin's point will be selected.

There are two types of selection: if the user selects one point, a Pushpin is displayed that the user can move around and delete. If the user selects multiple points (by selecting one point and then selecting another), the selection will display as a different colored line. Selection must be consecutive, so selecting two far away points will select all the points in between them on the path. The selection line can be moved and deleted as well, but not individual points within the line.

The user can double click on the map background to append a point to the path. The user can click on the Polyline (as long as the mouse is far away enough from the points as to not show the Nearby Pin), and this will insert a point in that segment.

To test this, loaded in a path with 10k points. Performance is good.

Added a MapItemClickDragBehavior class, which adds functionality to detect both clicks and drags on a MapItem. This is restricted to MapItems because it requires mouse events from the parent Map control.

The behavior has a click event that gets called when the user finishes clicking on the object (without moving the mouse). It has a drag event that gets called every time the user moves the mouse during a click event.

Note that because both MapItemClickDragBehavior and LeftButtonClickBehavior set MouseDown handled to true, they cannot be used together. This is okay, because MapItemClickDragBehavior implements the functionality of LeftButtonClickBehavior. LeftButtonClickBehavior is still useful because it does not require the component to be a Map item.

Switched over all the draggable pins and paths in both MapPath and MapLocation to use this behavior, including the MainMapItem in MapItemView