Fixed ListBox binding errors, hopefully forever
A found a fix for the binding errors on ListBoxItems' ContentAlignment properties. These binding errors were occuring when an event was moving between years/months. Perhaps the fix would have worked for all of these binding errors I was getting for so long.

The fix is to declare the ListBoxItem style property "OverridesDefaultStyle" as true.