Add map path/location, overhaul popup system
In the Map, Add Location and Add Path buttons now open popups to enter data about them. When creating a Path, the user can choose a txt file to load the path data from. Actual loading of the path data is not implemented yet.

When editing a path, if the user tries to leave edit mode and there are no points in the path, the user is now presented with a Y/N popup box that will delete the path. A path without points shouldn't exist.

Improved the popup window system. Now, the PopupWindow and PopupViewModel base classes handle the Accept vs Cancel buttons & interaction. When a popup is closed, the Navigator will automatically set the boolean to Accepted or Canceled in the PopupReturnArgs.

Also, data validation is now implemented. Subclasses of PopupViewModel must implement the ValidateData() method, which returns a bool which determines whether the user can press Accept or not. In this method, subclasses should set ValidationErrorText if any data fails, and the text will be displayed in red at the top of the popup window.

Added NotConverter, which takes the inverse of a boolean value.