Abstracted MediaCollection lists, improved FilterSet
MediaCollection was maintaing a list of all the tags in all of its children. Abstracted this functionality to a class called MaintainedParentCollection, that does the same thing for a generic collection and a generic field in the children. The collection class just has to give it a set of functions which determine what the field is in each child, whether its stored in a list or singly, etc.

Previously, if a MediaCollection has an Event as a child, it was subscribing to the tag changed events in each of the children of the event. It was then receiving those change events, as well as the duplicate change events when the event itself was changing its totaled tag list. So now, MaintainedParentCollection just subscribes to the collection changed event of the child, and not the child's children's events.

In FilterCriteria and FilterSet, when an item's filter criteria changes without knowing whether it tightens or loosens the filter, it makes more sense to have a separate FilterCriteriaChanged function rather than calling both FilterCriteriaTightened and FilterCriteriaLoosened. Also, in TimeRangeFilter, it wasn't even calling both of these, and this was a bug.